Cleaned up index management of all connections, publishers, and subscribers managed by publisher and subscriber contexts. Now uses uint_16 without negative numbers, using RMC_NIL_INDEX as an out-of-bounds value. Allows a uint32 to store the index in its lower 16 bits and additional data in the upper 16 bits.

+++ rmc_proto_test_common.c
@@ -28,7 +28,7 @@ void poll_add(user_data_t user_data,
     char buf[16];
     int epollfd = user_data.i32;
     struct epoll_event ev = {
-        .data.u32 = index,
+        .data.u32 = (uint32_t) index,
         .events = 0 // EPOLLONESHOT
     };
 
@@ -61,7 +61,7 @@ void poll_modify(user_data_t user_data,
 {
     int epollfd = user_data.i32;
     struct epoll_event ev = {
-        .data.u32 = index,
+        .data.u32 = (uint32_t) index,
         .events = 0 // EPOLLONESHOT
     };
